Add support for new Energy feature
The Energy panel, introduced in 2021.8, has some requirements for the sensor. This change sets the following for "Cumulative Active Import" and "Cumulative Active Export", allowing the sensor to be added to the Energy panel:
- `state_class` to `measurement`
- `device_class` to `energy`
- `last_reset_type` to `never`, which sets "epoch time".

Since Cumulative Active Export has the same unit I've made the same change there, but not sure if it is useful.

This should help with #12.